Updated test location
