--- main/integration-tests/__test__/api-tests/appstream-disabled/example-appstream-disabled.test.js
@@ -1,5 +1,0 @@
-describe('Example AppStream Disabled Tests', () => {
-  it('AppStream disabled', () => {
-    expect(true).toEqual(true);
-  });
-});
